Add timeout when stopping LDAPExploitServer in LDAP server test
The server is started in a daemon thread now, not a subprocess.
The timeout is needed so that it doesn't keep waiting for the thread
to terminate (the default timeout for joining the server thread is `None`).

Original commit https://github.com/guardicore/monkey/commit/6f72700c5514e1004e533abc23609e8e259f4e89